\name{nmgs_metapopulation_average}
\alias{nmgs_metapopulation_average}
\title{Average over the metapopulations.}
\usage{
nmgs_metapopulation_average(metacommunity)
}
\arguments{
  \item{metacommunity}{Returns:}
}
\value{
The average
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{utilities}

