% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nmgs_stats.R
\name{read_nmgs_stats}
\alias{read_nmgs_stats}
\title{Read in Simulation_out_s.csv}
\usage{
read_nmgs_stats(file)
}
\arguments{
\item{file}{NMGS output file name}
}
\value{
the statistics matrix (see above)
}
\description{
Gives statistics on the sampled communities.
}
\details{
Format:
SampleN,LN,LL,LO,HN,HL,HO,SN,SL,SO
where SampleN is nth sample generated under the neutral model with
fitted parameters taken from the corresponding MCMC sample.
LN,LL,LO are the log-likelihoods of the full neutral sample, the local
community sample and the observed sample respectively. Then HN,HL,HO
and SN,SL,SO are the species diversities (entropies) and richnesses.
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{utilities}

