% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{testf}
\alias{testf}
\title{testf
Test function}
\usage{
testf(x)
}
\arguments{
\item{x}{A number}
}
\value{
TBA
}
\description{
testf
Test function
}
\details{
TBA
}
\examples{
#  TBA
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\seealso{
TBA
}
\keyword{utilities}

