% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{nmgs_metapopulation_average}
\alias{nmgs_metapopulation_average}
\title{Average over the metapopulations.}
\usage{
nmgs_metapopulation_average(metacommunity)
}
\description{
Arguments:
  @param metacommunity
}
\details{
Returns:
  @return The average
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{utilities}

