% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/analysis.R
=======
% Please edit documentation in R/nmgs_posterior_summaries.R
>>>>>>> ae186b09b348f3f8c693c9677b0dcfcdc700095c
\name{nmgs_posterior_summaries}
\alias{nmgs_posterior_summaries}
\title{NMGS posterior summaries}
\usage{
nmgs_posterior_summaries(samples, burnin = 0, thinning = 1)
}
<<<<<<< HEAD
=======
\arguments{
\item{samples}{Posterior samples of the NMGS model (output from the read_nmgs function)}

\item{burnin}{Burn-in sample}

\item{thinning}{Thinning}
}
>>>>>>> ae186b09b348f3f8c693c9677b0dcfcdc700095c
\value{
median, mean, standard deviation and upper and lower 95\% confidence 
        intervals for the biodiversity parameter theta and the immigration rates. 
        Parameters x summaries matrix.
}
\description{
<<<<<<< HEAD
Arguments:
  @param samples Posterior samples of the NMGS model (output from the read_nmgs function)
  @param burnin Burn-in sample 
  @param thinning Thinning
}
\details{
Returns:

=======
: Compute summary statistics of the sampled MCMC parameters,
       excluding the burn-in samples and thinning.
}
\details{
>>>>>>> ae186b09b348f3f8c693c9677b0dcfcdc700095c
The confidence intervals are calculated as the 2.5\% and 97.5\% quantiles.
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{utilities}

