% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nmgs.R
\name{read_nmgs}
\alias{read_nmgs}
\title{Read in NMGS posterior samples}
\usage{
read_nmgs(file, burnin = 0)
}
\description{
Simulation_out.csv: Contains the MCMC sampled parameters in format:
iter1,theta,i1,..,iN
..
iterM,theta,i1,..,iN
where theta is the biodiversity parameter and i1,...,iN are the immigration rates to each sample

Arguments:
  @param file NMGS output file name

Returns:
  @return Posterior samples of the NMGS parameters
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{utilities}

