% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_nmgs_metacommunity}
\alias{read_nmgs_metacommunity}
\title{Read in NMGS generated metacommunities file}
\usage{
read_nmgs_metacommunity(file)
}
<<<<<<< HEAD
\description{
=======
\arguments{
\item{file}{NMGS output file name}
}
\value{
Generated metacommunities from the NMGS model. A list with the following elements:
    	     N: A matrix. Each row corresponds to a sample. SampleN,SL,SN for each sample (see above)
     p: metacommunity distribution (local assemply; see above)
     q: metacommunity distribution (sample generated under full neutral model; see above)
}
\description{
Simulation_out_m.csv gives the generated metacommunities with format:
  SampleN,SL,SN
  p1,...,pSL
  q1,...,qSN
>>>>>>> ae186b09b348f3f8c693c9677b0dcfcdc700095c
where SampleN is nth sample generated from the neutral model with
parameters from the corresponding MCMC sample. SL and SN are the
number of species in the local assembly sample (always S) and the
sample generated under the full neutral model respectively, p and q
<<<<<<< HEAD
are then the metacommunity distributions
}
\details{
Arguments:
  @param file NMGS output file name

Returns:
  @return Generated metacommunities from the NMGS model. A list with the following elements:
    	     N: A matrix. Each row corresponds to a sample. SampleN,SL,SN for each sample (see above)
     p: metacommunity distribution (local assemply; see above)
     q: metacommunity distribution (sample generated under full neutral model; see above)
=======
are then the metacommunity distributions.
>>>>>>> ae186b09b348f3f8c693c9677b0dcfcdc700095c
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{p1,...,pSL}
\keyword{q1,...,qSN}
\keyword{utilities}

