% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/io.R
\name{read_nmgs_stats}
\alias{read_nmgs_stats}
\title{Read in Simulation_out_s.csv: 
Gives statistics on the sampled communities in format:
SampleN,LN,LL,LO,HN,HL,HO,SN,SL,SO}
\usage{
read_nmgs_stats(file)
}
\description{
where SampleN is nth sample generated under the neutral model with
fitted parameters taken from the corresponding MCMC sample. LN,LL,LO
are the log-likelihoods of the full neutral sample, the local
community sample and the observed sample respectively. Then HN,HL,HO
and SN,SL,SO are the species entropies and richness's in the same
order.
}
\details{
Arguments:
  @param file NMGS output file name

Returns:
  @return the statistics matrix (see above)
=======
% Please edit documentation in R/read_nmgs_stats.R
\name{read_nmgs_stats}
\alias{read_nmgs_stats}
\title{Read in Simulation_out_s.csv}
\usage{
read_nmgs_stats(file)
}
\arguments{
\item{file}{NMGS output file name}
}
\value{
the statistics matrix (see above)
}
\description{
Gives statistics on the sampled communities.
}
\details{
Format:
SampleN,LN,LL,LO,HN,HL,HO,SN,SL,SO
where SampleN is nth sample generated under the neutral model with
fitted parameters taken from the corresponding MCMC sample.
LN,LL,LO are the log-likelihoods of the full neutral sample, the local
community sample and the observed sample respectively. Then HN,HL,HO
and SN,SL,SO are the species diversities (entropies) and richnesses.
>>>>>>> ae186b09b348f3f8c693c9677b0dcfcdc700095c
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("NMGS")
}
\keyword{utilities}

