\name{testf}
\alias{testf}
\title{testf
Test function}
\usage{
  testf(a, b)
}
\arguments{
  \item{a}{A number}

  \item{b}{Another Number}
}
\value{
  TBA
}
\description{
  testf Test function
}
\details{
  TBA
}
\examples{
#  TBA
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("NMGS")
}
\seealso{
  TBA
}
\keyword{utilities}

